+++ src/main/java/rldevs4j/base/env/gsmdp/evgen/ExogenousEventActivation.java
@@ -9,18 +9,18 @@ import java.util.Map;
  * @author Ezequiel Beccaría
  */
 public class ExogenousEventActivation extends entity{
-    private Map<String, Map<String,Double>> content;
+    private Map<String, Map<String,Float>> content;
 
-    public ExogenousEventActivation(Map<String, Map<String,Double>> content) {
+    public ExogenousEventActivation(Map<String, Map<String,Float>> content) {
         super("ExogenousEventGeneratorEntity");
         this.content = content;        
     }
 
-    public Map<String, Map<String, Double>> getContent() {
+    public Map<String, Map<String, Float>> getContent() {
         return content;
     }
 
-    public Map<String,Double> getIndividualContent(String generatorName){
+    public Map<String,Float> getIndividualContent(String generatorName){
         return content.get(generatorName);
     }
